Fix race
prev_offset is read by all worker threads and written by the thread with id 0 with the
assumption that the new value will not be used by the other threads. This is however not
enforced in any way and so worker threads with id != 0 might use the update value of prev_offset
erroneously.